Add messages module and remove bidding module
bidding files are included in allocation module
