+++ mrs/messages/bid.py
