+++ setup.py
@@ -5,7 +5,7 @@ from setuptools import setup
 setup(name='mrs',
       packages=['mrs', 'mrs.config', 'mrs.config.default',
                 'mrs.db.models', 'mrs.db.models.performance', 'mrs.exceptions',
-                'mrs.timetable', 'mrs.allocation', 'mrs.bidding', 'mrs.tests',
+                'mrs.timetable', 'mrs.allocation', 'mrs.tests',
                 'mrs.scheduling', 'mrs.dispatching', 'mrs.execution', 'mrs.utils'],
       version='0.2.0',
       description='Multi-Robot System (MRS) components for performing'
